Add pkg/ and move internal/buffer to pkg/bufferpool
